This EDP provides the guiding principles for migrating to cloud-based IT solutions for business owners and other VA stakeholders at OIT.
This EDP provides the guiding principles for migrating to cloud-based IT solutions for business owners and other VA stakeholders at OIT. The solutions entail access to virtual servers and storage, secure network connections, and well provisioned service models, such as Software-as-a-Service (SaaS), Platform-as-a-Service (PaaS), or Infrastructure-as-a-Service (IaaS). Cloud adoption is a business decision that will enhance VA’s mission, services, capabilities, and enable rapid development, deployment, and operations that comply with VA Handbook 6500 security and privacy guidelines.