This Enterprise Design Pattern focuses on the ECSB required to provide a cloud-based hosting environment for applications.
This Enterprise Design Pattern focuses on the ECSB required to provide a cloud-based hosting environment for applications. It will define the ECSB program for LOBs to leverage cloud services and outlines the standards and capabilities for an ECSB to act as a liaison between VA (Cloud Consumer) and Cloud Providers. This Enterprise Design Pattern will adhere to VA and federal policies.