This document provides an enterprise-level view of the “As-Is” and “To-Be” SOA capabilities within the VA IT infrastructure.
This document provides an enterprise-level view of the “As-Is” and “To-Be” SOA capabilities within the VA IT infrastructure. The document focuses on a vendor-independent framework for an enterprise SOA environment. It references, rather than duplicates, lower-level solution guidance. The focus of this EDP is on the background of the “As-Is” state of the VA SOA environment, descriptions of key components of the Enterprise SOA environment, the “To-Be” Enterprise SOA and associated attributes, and more.