This EDP focuses on the process for identifying, classifying, remediating, and mitigating vulnerabilities found within VA’s IT infrastructure that is integral to both computer and network security.
This EDP focuses on the process for identifying, classifying, remediating, and mitigating vulnerabilities found within VA’s IT infrastructure that is integral to both computer and network security. A robust VM program should be implemented within the enterprise in order to effectively remedy vulnerabilities identified in operating system (OS) databases, applications, and other network devices. This EDP makes recommendations for implementing standardized enterprise-wide VM practices that are based on industry best practices.