This Enterprise Design Pattern provides strategic direction for the VA to establish a capability for standardized access for multiple applications to enterprise data stores.
This Enterprise Design Pattern provides strategic direction for the VA to establish a capability for standardized access for multiple applications to enterprise data stores. In the target VA IT environment, enterprise storage, retrieval, and exchange of data will be achieved through the reuse of Enterprise Shared Services (ESS) provided by the VA’s SOA infrastructure, which includes the Enterprise Messaging Infrastructure (eMI).